pyproject.toml
[tool.poetry]
name = "secrethitlerstatbot"
version = "0.1.0"
description = "Telegram bot for recording Secret Hitler statistic"
authors = ["alex"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
aiosqlite = "^0.19.0"
python-dotenv = "^1.0.1"
python-telegram-bot = "^20.7"
pydantic = "^2.5.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
